Abstract assessment section presenter
Both risk and healtchare assessment section presenters follow the same
patterns. This work abstract that behaviour into a single section
presenter that is inherited by the risk and healthcare presenters.
